requirements.txt
streamlit>=1.41.0
pandas>=2.2.3
plotly>=5.24.1
requests>=2.32.3
python-dotenv>=1.0.1
pytest>=8.3.3
